requirements.txt
ipykernel==6.29.4
langchain-core~=0.3
langchain~=0.3
langchain_community~=0.3
langchain_openai~=0.3
langchain-prompty~=0.1.1
python-dotenv==1.0.1
pyplz~=0.11.1
azure-identity==1.20.0
pydantic==2.8.2